requirements.txt
django
django-filter
yfinance
-e git+https://github.com/anurag3753/tradewise#egg=tradewise